+++ 4_abcTS/vite.config.js
@@ -11,6 +11,8 @@ export default defineConfig({
     open: '/editor.html' // 👈 啟動時自動在瀏覽器打開 editor.html
   },
   build: {
+    // 1. 💡 關閉壓縮 (預設是 'esbuild')
+    minify: false,
     // 禁用 CSS 代碼拆分，讓樣式跟著打包，或獨立為單一 css
     cssCodeSplit: false,
     lib: {